tables: 1.13.00 update — all V3-apply tables parse (iteminfo + 8 tables)
Game 1.12.2 -> 1.13.00 (0.paver 01000d00...). V3 harness: 130 OK / 0 FAIL
(all mod-apply tables parse). Verified byte-exact roundtrip on fresh vanilla
fixtures (2026-7-3), added to testenv FALLBACK_DIRS.

- iteminfo: SubItem disc 17 zero-payload (all 4 arms); PrefabData unified with
  the old gimmick_visual_prefab_data_list (adds scale/animation_path_list/
  use_gimmick_prefab/prefab_data_type); prefab_data_list relocated to the record
  tail + 2 trailing bytes. Field order from the game's IDA readers; full-table
  byte-exact roundtrip over all 6508 items.
- equip_slot_info: +7B/EquipInfoData (u32+u16+u8; u8 = _isHideEquipInDyeingProcess).
- part_prefab_dye_slot_info: unk_u32_112 was the _modelPropertyDyeInfoList count;
  decoded as CArray<DyeTextureSet> (3 CString + 4 u8) — fixes dyeable weapons/disguises.
- vehicle_info +4 trailing u32; house_info +2/region elem; bank_info +27 (u32/u32/u8/CString);
  quick_slot_info show_item_group_key u16->u8; contents_phase_info +2 mid + 17B tail.

Known gap: 8 tables still fail STRICT roundtrip but blob-fall-back in V3
(effect/region/store/game_play_trigger/gimmick_group/sequencer_spawn/
spawning_pool_auto_spawn/terrain_region_auto_spawn) — pending typed decode.
inventory verified clean (variable list growth only, no drift).

+++ src/tables/quick_slot_info/info.rs
@@ -20,7 +20,10 @@ py_binary_struct! {
         pub item_info_quick_slot_index: u8,
         pub use_special_reserve_slot: u8,
         pub use_on_select: u8,
-        pub show_item_group_key: u16,
+        // 1.13.00: narrowed u16 -> u8 (−1 B per QuickSlotItem). Byte-diff
+        // decisive: exactly one byte removed at element-offset +40 (this field's
+        // high byte) in every slot element; all following fields shift −1.
+        pub show_item_group_key: u8,
         pub show_item_tag: u32,
         pub fixed_status_key: u32,
         pub select_special_name: CString<'a>,
